SinglyLinkedList upload
An SLList nearly implements the full set of Deque operations. The only missing operation is removing from the tail of an SLList. 
Removing from the tail of an SLList is difficult because it requires updating the value of tail so that it points to the node w that precedes tail in the SLList; this is the node w such that w.next=tail. 
Unfortunately, the only way to get to w is by traversing the SLList starting at head and taking n-2 steps.